approach2/requirements.txt
a2a-sdk==0.3.8
agent-framework==1.0.0b251007
agent-framework-a2a==1.0.0b251007
agent-framework-azure-ai==1.0.0b251007
agent-framework-copilotstudio==1.0.0b251007
agent-framework-core==1.0.0b251007
agent-framework-devui==1.0.0b251007
agent-framework-mem0==1.0.0b251007
agent-framework-redis==1.0.0b251007
aiofiles==25.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.13.0
aiosignal==1.4.0
alembic==1.17.0
annotated-types==0.7.0
anthropic==0.71.0
anyio==4.11.0
asgiref==3.10.0
asttokens==3.0.0
attrs==25.4.0
Authlib==1.6.5
azure-ai-agents==1.2.0b5
azure-ai-projects==1.1.0b4
azure-core==1.36.0
azure-core-tracing-opentelemetry==1.0.0b12
azure-identity==1.25.1
azure-monitor-opentelemetry==1.8.1
azure-monitor-opentelemetry-exporter==1.0.0b44
azure-storage-blob==12.27.0
backoff==2.2.1
blinker==1.9.0
boto3==1.40.59
botocore==1.40.59
cachetools==6.2.1
certifi==2025.10.5
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.0
cloudpickle==3.1.1
colorama==0.4.6
comm==0.2.3
contourpy==1.3.3
cryptography==45.0.7
cycler==0.12.1
cyclopts==3.24.0
databricks-sdk==0.68.0
debugpy==1.8.17
decorator==5.2.1
distro==1.9.0
dnspython==2.8.0
docker==7.1.0
docstring_parser==0.17.0
docutils==0.22.2
email-validator==2.3.0
exceptiongroup==1.3.0
executing==2.2.1
faiss-cpu==1.12.0
fastapi==0.119.0
fastmcp==2.12.4
fixedint==0.1.6
Flask==3.1.2
flask-cors==6.0.1
fonttools==4.60.1
frozenlist==1.8.0
gitdb==4.0.12
GitPython==3.1.45
google-api-core==2.26.0
google-auth==2.41.1
googleapis-common-protos==1.70.0
graphene==3.4.3
graphql-core==3.2.6
graphql-relay==3.2.0
greenlet==3.2.4
grpcio==1.75.1
h11==0.16.0
h2==4.3.0
hpack==4.1.0
httpcore==1.0.9
httptools==0.7.1
httpx==0.28.1
httpx-sse==0.4.3
hubspot-api-client==12.0.0
hyperframe==6.1.0
idna==3.11
importlib_metadata==8.7.0
ipykernel==7.0.1
ipython==9.6.0
ipython_pygments_lexers==1.1.1
isodate==0.7.2
itsdangerous==2.2.0
jedi==0.19.2
Jinja2==3.1.6
jiter==0.11.0
jmespath==1.0.1
joblib==1.5.2
jsonpatch==1.33
jsonpath-ng==1.7.0
jsonpointer==3.0.0
jsonschema==4.25.1
jsonschema-path==0.3.4
jsonschema-specifications==2025.9.1
jupyter_client==8.6.3
jupyter_core==5.9.1
kiwisolver==1.4.9
langchain==1.0.2
langchain-core==1.0.1
langgraph==1.0.1
langgraph-checkpoint==3.0.0
langgraph-prebuilt==1.0.1
langgraph-sdk==0.2.9
langsmith==0.4.38
lazy-object-proxy==1.12.0
Mako==1.3.10
markdown-it-py==4.0.0
MarkupSafe==3.0.3
matplotlib==3.10.7
matplotlib-inline==0.1.7
mcp==1.17.0
mdurl==0.1.2
mem0ai==0.1.118
microsoft-agents-activity==0.4.0
microsoft-agents-copilotstudio-client==0.4.0
microsoft-agents-hosting-core==0.4.0
ml_dtypes==0.5.3
mlflow==3.4.0
mlflow-skinny==3.4.0
mlflow-tracing==3.4.0
more-itertools==10.8.0
msal==1.34.0
msal-extensions==1.3.1
msrest==0.7.1
multidict==6.7.0
nest-asyncio==1.6.0
numpy==2.3.3
oauthlib==3.3.1
openai==1.109.1
openapi-core==0.19.5
openapi-pydantic==0.5.1
openapi-schema-validator==0.6.3
openapi-spec-validator==0.7.2
opentelemetry-api==1.37.0
opentelemetry-exporter-otlp-proto-common==1.37.0
opentelemetry-exporter-otlp-proto-grpc==1.37.0
opentelemetry-instrumentation==0.58b0
opentelemetry-instrumentation-asgi==0.58b0
opentelemetry-instrumentation-dbapi==0.58b0
opentelemetry-instrumentation-django==0.58b0
opentelemetry-instrumentation-fastapi==0.58b0
opentelemetry-instrumentation-flask==0.58b0
opentelemetry-instrumentation-psycopg2==0.58b0
opentelemetry-instrumentation-requests==0.58b0
opentelemetry-instrumentation-urllib==0.58b0
opentelemetry-instrumentation-urllib3==0.58b0
opentelemetry-instrumentation-wsgi==0.58b0
opentelemetry-proto==1.37.0
opentelemetry-resource-detector-azure==0.1.5
opentelemetry-sdk==1.37.0
opentelemetry-semantic-conventions==0.58b0
opentelemetry-semantic-conventions-ai==0.4.13
opentelemetry-util-http==0.58b0
orjson==3.11.4
ormsgpack==1.11.0
packaging==25.0
pandas==2.3.3
parse==1.20.2
parso==0.8.5
pathable==0.4.4
pillow==11.3.0
platformdirs==4.5.0
ply==3.11
portalocker==3.2.0
posthog==6.7.7
prompt_toolkit==3.0.52
propcache==0.4.1
proto-plus==1.26.1
protobuf==5.29.5
psutil==7.1.0
psycopg2==2.9.11
psycopg2-binary==2.9.11
pure_eval==0.2.3
pyarrow==21.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.23
pydantic==2.12.2
pydantic-settings==2.11.0
pydantic_core==2.41.4
Pygments==2.19.2
PyJWT==2.10.1
pyparsing==3.2.5
pyperclip==1.11.0
python-a2a==0.5.10
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python-multipart==0.0.20
python-ulid==3.1.0
pytz==2025.2
pywin32==311
PyYAML==6.0.3
pyzmq==27.1.0
qdrant-client==1.15.1
redis==6.4.0
redisvl==0.9.1
referencing==0.36.2
requests==2.32.5
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
rfc3339-validator==0.1.4
rich==14.2.0
rich-rst==1.3.2
rpds-py==0.27.1
rsa==4.9.1
s3transfer==0.14.0
scikit-learn==1.7.2
scipy==1.16.2
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
SQLAlchemy==2.0.44
sqlparse==0.5.3
sse-starlette==3.0.2
stack-data==0.6.3
starlette==0.48.0
tenacity==9.1.2
threadpoolctl==3.6.0
tornado==6.5.2
tqdm==4.67.1
traitlets==5.14.3
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
uvicorn==0.37.0
waitress==3.0.2
watchfiles==1.1.1
wcwidth==0.2.14
websockets==15.0.1
Werkzeug==3.1.1
wrapt==1.17.3
xxhash==3.6.0
yarl==1.22.0
zipp==3.23.0
zstandard==0.25.0